+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Beta 0.5.3] 2025-02-14
+
+### Changed
+- /calendar and /clrecord wont send files, just go check google drive.
+
+### Fixed
+- Clockin record's output.
+
+---
 ## [Beta 0.5.2] 2025-02-14
 
 ### Changed